requirements.txt
asgiref==3.5.2
Django==4.1.1
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
gunicorn==20.1.0
